# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API - Serviceasy",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "User",
      "description": "Endpoints User"
    },
    {
      "name": "Auth",
      "description": "Endpoints Auth"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for user register.",
        "parameters": [
          {
            "name": "Register",
            "in": "body",
            "description": "Data for register",
            "schema": {
              "$ref": "#/definitions/DataUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for view data of user.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for update data for user.",
        "parameters": [
          {
            "name": "Update",
            "in": "body",
            "description": "Data for update register",
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for delete user.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Login",
        "parameters": [
          {
            "name": "Login",
            "in": "body",
            "description": "Provide email and password registred",
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/verifyCode": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Checking verification code",
        "parameters": [
          {
            "name": "Checking",
            "in": "body",
            "description": "Provide email and verification code",
            "schema": {
              "$ref": "#/definitions/VerifyUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/forgotPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Request verification code for update password",
        "parameters": [
          {
            "name": "Verification",
            "in": "body",
            "description": "Provide email registered",
            "schema": {
              "$ref": "#/definitions/ForgotPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/updatePassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Update password",
        "parameters": [
          {
            "name": "Update",
            "in": "body",
            "description": "Provide email and new password",
            "schema": {
              "$ref": "#/definitions/UpdatePassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "DataUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "laurammoraes2@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "Teste-123"
        },
        "phoneNumber": {
          "type": "string",
          "example": "+5519993417478"
        }
      },
      "required": [
        "email",
        "password",
        "phoneNumber"
      ]
    },
    "VerifyUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "laurammoraes2@gmail.com"
        },
        "code": {
          "type": "string",
          "example": "XXXX"
        }
      },
      "required": [
        "email",
        "code"
      ]
    },
    "LoginUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "laurammoraes2@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "Ritinha-09"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "ForgotPassword": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "laurammoraes2@gmail.com"
        }
      },
      "required": [
        "email"
      ]
    },
    "UpdatePassword": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "laurammoraes2@gmail.com"
        },
        "code": {
          "type": "string",
          "example": "733404"
        },
        "newPassword": {
          "type": "string",
          "example": "Laura-123"
        }
      },
      "required": [
        "email",
        "code",
        "newPassword"
      ]
    },
    "UpdateUser": {
      "type": "object",
      "properties": {
        "newPhoneNumber": {
          "type": "string",
          "example": "+5519993419999"
        }
      },
      "required": [
        "newPhoneNumber"
      ]
    }
  }
}